Rename camtool "schema" to "prefab"
